propagate_tensor_shape调整到可以使用，有很多地方被强制变成了浮点（例如发射的脉冲）

+++ paibox/fx_converter/core_op.py
@@ -209,7 +209,7 @@ class CoreOpNode(nn.Module, PAIIR):
     def forward(self, *xs: torch.Tensor) -> torch.Tensor:
         s = 0
         for sign, op, x in zip(self.signs, self.op1, xs, strict=True):
-            output = op(x)
+            output = op(x.float())
             s += sign * output
 
         return self.op2(s)
